Make migration 0009 pass in non-dry mode

+++ facebook_auth/migrations/0009_auto__add_unique_usertoken_token.py
@@ -8,8 +8,9 @@ from django.db import models
 class Migration(SchemaMigration):
 
     def forwards(self, orm):
-        for token in orm['facebook_auth.UserToken'].objects.all():
-            orm['facebook_auth.UserToken'].objects.filter(token=token.token, id__gt=token.id).delete()
+        if not db.dry_run:
+            for token in orm['facebook_auth.UserToken'].objects.all():
+                orm['facebook_auth.UserToken'].objects.filter(token=token.token, id__gt=token.id).delete()
         db.commit_transaction()
         db.start_transaction()
         # Adding unique constraint on 'UserToken', fields ['token']
